Statepoints: Track Basepointers only if Required by the Runtime
This change parametrizes the RewriteStatepointsForGC phase on whether
base-pointers should be explicitly tracked for each live GC-pointer.

Runtimes like the CLR do not require reporting the base-pointers for each
live managed pointer at gc-safepoints. When base-pointer tracking is
unnecessary, the gc-pointer -> base-pointer mapping is treated as identity.

This change does not affect code generation in LLVM tests, as there are
currently no clients for the configuration where base-pointers are not tracked.