Initial workup for alloca and inline stack probes for CoreCLR.
Add the notion of a CoreCLR environment and use it to drive a different expansion of alloca.

Still much left undone: the actual inline probe sequence, handling the case where the probe is needed in the prolog, x86 support, non-windows support, and optimizations when we know we're going to be zeroing the allocated memory.