Add an intrinsic for calls that are GC transitions.
This intrinsic may be used to support calls from functions that are
GC-aware ("managed functions") to functions that are not GC-aware
("unmanaged functions"). This intrinsic is similar to the existing
statepoint instruction, except that it offers the backend the
opportunity to insert (somewhat) arbitrary code to managed the
transition from managed code to unmanaged code and back.

In order to support the injection of transition code, this intrinsic
wraps the STATEPOINT DAG node generated by the usual statepoint
lowering with two additional nodes: GC_TRANSITION_START and
GC_TRANSITION_END. The intrinsic accepts an additional set of
parameters that are supplied as operands to these nodes that may
be used by the backend during code generation.