Implement the prolog inline stack probe.
This change enables use of the inline stack probe within the method prolog, for cases where the local frame size exceeds 4096 bytes. The existing alloca code is extended to handle the prolog case by allowing precoloring of the various virtual registers to RCX and RDX, saving/restoring RCX/RDX around the entire probe sequence. For now the save/restore is done via push/pop; if this proves to be not viable long term we can try and get them into the caller-supplied save area.

Upstream frame-gen logic is generalized to handle the case where the in-prolog check creates control flow and hence changes the "current" basic block. Responsibilty for marking the added MachineInstrs as frame setup is moved to the expansions themselves.