Support CoreCLR filters
Expand HandlerTreeNode to have a ClauseKind (now that there are three) and
a FilterFunction used in filter handlers.

Change ClrEHClause to pass along a HandlerTreeNode opaquely rather than
replicating its fields.

Ensure that outlined filter handlers are inserted in the function list
just after the filters they handle.

Check in a few places that are determining whether to emit EH-relevant
info whether a function has a Win EH parent, so that early-outlined
filters are included.