Avoid phi update if instruction list is empty.
Ran into an issue that looked like uninitialized data when the phi update code
was called on a successor that containes no instructions in SplitBlock.  Case
arises in LLILC durring flow graph build.  The flow graph build algorithm
first constructs all the empty blocks, then fills them out as the MSIL is
parsed.  Some blocks with unexpected branch targets need to be split at this
time which can cause the phi update code to be run on an empty block - and
uninitialized data looks enough like a phi that bad things happen to the flow
graph.