[CodeGen] Mark setjmp/catchret MBBs address-taken
Doing so ensures that BranchFolding (and similar) won't remove them.

Also allow AsmPrinter::EmitBasicBlockStart to process MBBs which are
address-taken but do not have BBs that are address-taken, since otherwise
its call to getAddrLabelSymbolTableToEmit would fail an assertion on such
blocks.  I audited the other callers of getAddrLabelSymbolTableToEmit;
they all have BBs known to be address-taken except for the call through
getAddrLabelSymbol from WinException::create32bitRef; that call is
actually unreachable, so I've removed it and updated the signature of
create32bitRef.

This fixes PR25168.