Disable Function-entry polls for CoreCLR
Disble safepoint poll()s at function-entry for CoreCLR, in order to
work around a problem in LLILC, where the safepoint poll is introduced
before GC-aggregates are initialized by a memset() call in the prolog.
Issue #892

In the long term, CoreCLR does not need Function-entry safepoints
because of return-address hijacking. But, we need to insert
safepoint-polls at tail calls Issue #425