Explicitly pass InvokeInst to lowerInvokable
The CallLoweringInfo passed to lowerInvokable for statepoint calls does
not have the CallSite associated with it, but when lowerInvokable gets
passed a funclet-style EHPadBB it needs to find the InvokeInst so that it
can look up the state number in the InvokeStateMap.  The lowerInvokable
method is a private helper, so add an explicit InvokeInst parameter to it
and update both callers to set it appropriately.